+++ src/main/java/com/andersmmg/falloutstuff/block/entity/VaultSignBlockEntity.java
@@ -5,8 +5,12 @@ import com.andersmmg.falloutstuff.record.SignUpdatePacket;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.nbt.NbtCompound;
+import net.minecraft.network.listener.ClientPlayPacketListener;
+import net.minecraft.network.packet.Packet;
+import net.minecraft.network.packet.s2c.play.BlockEntityUpdateS2CPacket;
 import net.minecraft.text.Text;
 import net.minecraft.util.math.BlockPos;
+import org.jetbrains.annotations.Nullable;
 
 public class VaultSignBlockEntity extends BlockEntity {
     private Text text = Text.literal("");
@@ -46,6 +50,12 @@ public class VaultSignBlockEntity extends BlockEntity {
     // Method to send update packet to clients
     private void sendUpdatePacket() {
         FalloutStuff.SIGN_UPDATE_CHANNEL.clientHandle().send(new SignUpdatePacket(pos, this.text));
+    }
+
+    @Nullable
+    @Override
+    public Packet<ClientPlayPacketListener> toUpdatePacket() {
+        return BlockEntityUpdateS2CPacket.create(this);
     }
 
     @Override